requirements.txt
python-dotenv>=1.0.0
python-telegram-bot[job-queue]==21.6
supabase>=2.9.1
aiohttp>=3.10.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
